backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
sqlalchemy
pymysql
psycopg2-binary
langchain
langchain-community
python-dotenv